feat(default): move fn to default
move the two function to types mod
- get_url_resource
- get_url_resource2
