+++ src/types.rs
@@ -20,6 +20,7 @@ mod test;
 
 #[cfg(feature = "core")]
 pub use self::core::{ContentRange, Query, QueryKey, QueryValue, SetOssQuery};
+use self::object::{ObjectPathInner, SetObjectPath};
 
 const OSS_DOMAIN_PREFIX: &str = "https://oss-";
 #[allow(dead_code)]
@@ -1171,4 +1172,38 @@ impl PartialEq<InnerCanonicalizedResource<'_>> for &str {
     fn eq(&self, other: &InnerCanonicalizedResource<'_>) -> bool {
         self == &other.0
     }
+}
+
+/// 根据 endpoint， bucket， path 获取接口信息
+pub fn get_url_resource(
+    endpoint: &EndPoint,
+    bucket: &BucketName,
+    path: &ObjectPathInner,
+) -> (Url, CanonicalizedResource) {
+    let mut url = url_from_bucket(endpoint, bucket);
+    url.set_object_path(path);
+
+    let resource = CanonicalizedResource::from_object((bucket.as_ref(), path.as_ref()), []);
+
+    (url, resource)
+}
+
+/// 根据 endpoint， bucket， path 获取接口信息
+pub fn get_url_resource2<E: AsRef<EndPoint>, B: AsRef<BucketName>>(
+    endpoint: E,
+    bucket: B,
+    path: &ObjectPathInner,
+) -> (Url, CanonicalizedResource) {
+    get_url_resource(endpoint.as_ref(), bucket.as_ref(), path)
+}
+
+pub(crate) fn url_from_bucket(endpoint: &EndPoint, bucket: &BucketName) -> Url {
+    let url = format!(
+        "https://{}.oss-{}.aliyuncs.com",
+        bucket.as_ref(),
+        endpoint.as_ref()
+    );
+    url.parse().unwrap_or_else(|_| {
+        unreachable!("covert to url failed, bucket: {bucket}, endpoint: {endpoint}")
+    })
 }
